+++ models/restaurant.ts
@@ -2,7 +2,7 @@ import { sequelize, Sequelize } from '../src/lib/sequelize';
 
 const restaurant = sequelize.define('restaurant', {
   name: Sequelize.STRING,
-  cashBalance: Sequelize.DECIMAL
+  cashBalance: Sequelize.FLOAT
 });
 
 export default restaurant;
