+++ models/user.ts
@@ -2,7 +2,7 @@ import { sequelize, Sequelize } from '../src/lib/sequelize';
 
 const user = sequelize.define('user', {
   name: Sequelize.STRING,
-  cashBalance: Sequelize.DECIMAL
+  cashBalance: Sequelize.FLOAT
 });
 
 export default user;
